requirements.txt
asgiref==3.6.0
Django==4.1.6
django-admin-tools==0.9.2
django-appconf==1.0.5
django-select2==8.1.1
django-suit==0.2.28
Pillow==9.4.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2022.7
